requirements.txt
requests==2.32.3
pytest==8.2.0
allure-pytest==2.13.5
